requirements.txt
numpy==2.2.6
pandas==2.3.3
matplotlib==3.10.8
japanize-matplotlib==1.1.3
transformers==4.57.6
datasets==4.6.1
scikit-learn==1.7.2
pillow==12.0.0
opencv-python
tqdm==4.67.3
wandb==0.25.0
jupyterlab==4.5.5
ipykernel==7.2.0
pytorch-fid==0.3.0
